pyproject.toml
[build-system]
requires = ["setuptools >= 40.9.0", 'setuptools_scm[toml]>=6.2']
build-backend = "setuptools.build_meta"


[project]
name = "nskit"
authors = [{name = "djpugh@gmail.com", email="djpugh@gmail.com"}]
maintainers = [{ name = "djpugh@gmail.com", email="djpugh@gmail.com" }]
description = "Namespace Development Kit, providing a set of useful utilities for building software (in python namespaces)"
readme = {file="README.md", content-type = "text/x-rst"}
keywords = []
requires-python = '>=3.8'
classifiers = [
        "Environment :: Console",
        "Topic :: Software Development :: Libraries :: Python Modules",
        "Programming Language :: Python :: 3",
        "Programming Language :: Python :: 3.7",
        "Programming Language :: Python :: 3.8",
        "Programming Language :: Python :: 3.9",
        "Programming Language :: Python :: 3.10",
        "Programming Language :: Python :: 3.11",
        "Intended Audience :: Developers",
    	"License :: OSI Approved :: MIT License",
        "Operating System :: Microsoft :: Windows",
        "Operating System :: POSIX :: Linux",
        "Operating System :: MacOS :: MacOS X",
        "Natural Language :: English"
        # "Development Status :: 6 - Mature"
]
dependencies = [
    "aenum",
    "azure_devops_artifacts_helpers",
    "gitpython",
    "jinja2",
    "logzero",
    "orjson",
    "pydantic[email]",
    "pydantic-settings",
    "pygithub",
    "python-dotenv",
    "ruamel.yaml",
    "tomlkit",
    'typing-extensions; python_version < "3.12"',
    "virtualenv",
]
dynamic = ["version"]


[project.optional-dependencies]
azure_devops = [
    "azure-cli"
]
github = [
    "PyGithub"
]
dev = [
    "setuptools_scm[toml]",
    'importlib_metadata; python_version < "3.8"',
    "nox",
    'tomli; python_version < "3.11"',
    "pre-commit"
]
dev-lint = [
    "flake8>=5.0.4",
    "flake8-import-order>=0.18.2",
    "flake8-eradicate>=1.4.0",
    "flake8-docstrings>=0.2.7",
    "flake8-pyproject>=1.2.3",
    "flake8-bugbear>=22.0.0",
    "flake8-comprehensions>=3.12.0",
    "flake8-unused-arguments>=0.0.13",
    "flake8-noqa>=1.3.1"
]
dev-security = [
    "pipenv",
    "bandit"
]
dev-types = [
    "mypy",
    "lxml"
]
dev-test= [
    "pytest>=7.3.1",
    "pytest-cov>=4",
    "pytest-subtests"
    ]
dev-docs = [
    "mkdocs",
    "mkdocstrings[python]",
    "mkdocs-git-revision-date-localized-plugin",
    "mkdocs-material",
    "mkdocs-git-authors-plugin",
    "mkdocs-coverage",
    # "mkdocs-gen-files",
    # "mkdocs-literate-nav",
    "mike",
    "sphinx_github_changelog"
]
dev-build = [
    "build"
]

[project.urls]
repository = "https://github.com/djpugh/nskit"

[project.scripts]

[project.entry-points."nskit.recipes"]
python_package = "nskit.recipes.python.package:PackageRecipe"
python_api_service = "nskit.recipes.python.api:APIRecipe"
recipe = "nskit.recipes.recipe:RecipeRecipe"

[project.entry-points."nskit.vcs.providers"]
azure_devops = "nskit.vcs.providers.azure_devops:AzureDevOpsSettings"
github = "nskit.vcs.providers.github:GithubSettings"

# Tools
[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = [
    "*.md",
    "*.ini",
    "*.toml",
    "*.whl",
    "*.template",
    "*.bicep"
]

[tool.setuptools_scm]
write_to ="src/nskit/_version.py"
write_to_template = '__version__ = "{version}"'
version_scheme = "no-guess-dev"

[tool.mypy]
python_version = "3.11"
show_error_codes = true
strict = true
overrides = [
  { module = [
    "virtualenv.*",
    "tox.*",
    "importlib_metadata.*",
    "nskit._version"
  ], ignore_missing_imports = true },
]

[tool.flake8]
max-complexity = 22
max-line-length = 120
unused-arguments-ignore-abstract-functions = true
noqa-require-code = true
dictionaries = ["en_US", "python", "technical", "django"]
ignore = [
  "E203", # whitespace before ':'
  "W503", #  line break before binary operator
  "E501",
  "E226",
  "E121",
  "E123",
  "E126",
  "E24",
  "E704"
]
count = true
exclude = [
    # No need to traverse our git directory
    ".git",
    "./.git",
    # There's no value in checking cache directories
    "__pycache__",
    # The conf file is mostly autogenerated, ignore it
    "docs/source/conf.py",
    "./docs/source/conf.py",
    # This contains built code
    "build/*",
    "./build/lib/*",
    "./build/*",
    # This contains the builds that we don't want to check
    "dist",
    # This contains the docs
    "docs/*",
    # Compiled files
    "*.pyc",
    # Docs files
    "*.rst",
    "*.md",
    "*.png",
    ".tox",
    "examples",
    "tests",
    ".venv/*",
    ".tox/*",
    "src/nskit/_version.py",
    ".venv*/*"

]
show_source = true
statistics = true
application-import-names="nskit"
application-package-names="nskit"
docstring-convention="google"
import-order-style="appnexus"

[tool.isort]
profile = "appnexus"
src_paths = ['src/']
known_first_party = [
]

known_application = 'nskit*'
force_alphabetical_sort_within_sections = true
force_sort_within_sections = true
reverse_relative = true